pyproject.toml
[tool.poetry]
name = "league-planner"
version = "0.0.0"
description = ""
authors = ["Piotr Karaś <pepe500p@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
DateTime = "^5.2"
Django = "^4.2.6"
djangorestframework = "^3.14.0"
django-environ = "^0.4.5"
django-filter = "^23.3"
dumb-init = "^1.2.5.post1"
pillow = "^10.1.0"
poetry-dynamic-versioning = "^0.25.0"
psycopg2-binary = "^2.8.4"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
dotenv-linter = "^0.4.0"
factory-boy = "^3.3.0"
fawltydeps = "^0.13.1"
flake8 = "^6.0.0"
flake8-bandit = "^4.1.1"
flake8-bugbear = "^23.2.13"
flake8-builtins = "^2.1.0"
flake8-commas = "^2.1.0"
flake8-comprehensions = "^3.10.1"
flake8-debugger = "^4.1.2"
flake8-isort = "^6.0.0"
flake8-print = "^5.0.0"
flake8-pyproject = "^1.2.2"
flake8-pytest = "^1.4"
flake8-pytest-style = "^1.7.2"
ipython = "^8.14.0"
isort = {version = "^5.12.0", extras = ["colors"]}
mypy = "^1.0.1"
pytest = "^7.4.2"
pytest-cov = "^4.0.0"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.1"

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning"]
build-backend = "poetry_dynamic_versioning.backend"

[tool.black]
line-length = 120

# coverage settings
[tool.coverage.run]
branch = true
omit = [
    "*urls*",
    "*test*",
    "*admin*",
    "*__init__*",
    "./app/routes/*",
    "/usr/local/src/*",
    "*.venv*",
    "get-pip.py",
]

[tool.fawltydeps]
code = ["src"]
deps = ["pyproject.toml"]
ignore_unused = [
    # used indirectly
    "DateTime",
    "ipython",
    "pillow",  # ImageField
    "poetry-dynamic-versioning",
    "psycopg2-binary",
    "pytest-cov",

    # used in Dockerfile
    "dumb-init",
    "uvicorn",

    # Dev dependencies. This list should contain ALL of them!
    "black",
    "dotenv-linter",
    "fawltydeps",
    "flake8",
    "flake8-bandit",
    "flake8-bugbear",
    "flake8-builtins",
    "flake8-commas",
    "flake8-comprehensions",
    "flake8-debugger",
    "flake8-isort",
    "flake8-print",
    "flake8-pyproject",
    "flake8-pytest",
    "flake8-pytest-style",
    "isort",
    "mypy",
    "pytest",
    "pytest-asyncio",
]

[tool.flake8]
# S101 == Use of assert detected. Conflicts with pytest.
# W503 == Line break before binary operator. Conflicts with W504 - Line break after binary operator.
ignore = ["S101", "W503"]
max-line-length = 120

[tool.isort]
line_length = 120
order_by_type = false
profile = "black"

[tool.mypy]
disallow_untyped_defs = true
ignore_missing_imports = true
no_implicit_optional = false
disable_error_code = "misc"
exclude = ["/migrations/", "/tests/"]

# pytest settings
[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "league_planner.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = "--strict-markers -p no:warnings --cov=. --cov-fail-under=85 --cov-config=.coveragerc"
markers = [
    "integration: marks tests as integration (deselect with '-m \"not integration\"' to have just a quick check)",
    "serial",
    "print_queries: prints tests queries",
]
asyncio_mode = "auto"
